requirements.txt
pyrebase4==4.6.0
scikit-learn==1.0.2
keras==2.9.0
scipy==1.7.3
tensorflow
streamlit-authenticator==0.2.1
pyyaml
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
httpx_oauth
asyncio